✨ feat : add admin authentication feature

+++ backend/src/main/java/org/anotherclass/colortherock/domain/report/controller/AdminReportController.java
@@ -5,11 +5,14 @@ import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.anotherclass.colortherock.domain.member.entity.AdminDetails;
 import org.anotherclass.colortherock.domain.report.response.AdminReportDetailResponse;
 import org.anotherclass.colortherock.domain.report.response.AdminReportedPostResponse;
 import org.anotherclass.colortherock.domain.report.service.AdminReportService;
 import org.anotherclass.colortherock.global.common.BaseResponse;
+import org.anotherclass.colortherock.global.error.GlobalBaseException;
 import org.anotherclass.colortherock.global.error.GlobalErrorCode;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -27,7 +30,10 @@ public class AdminReportController {
     @ApiResponses({
             @ApiResponse(responseCode = "200", description = "영상게시글 불러오기 성공")
     })
-    public BaseResponse<List<AdminReportedPostResponse>> getReportList() {
+    public BaseResponse<List<AdminReportedPostResponse>> getReportList(@AuthenticationPrincipal AdminDetails adminDetails) {
+        if (adminDetails == null) {
+            throw new GlobalBaseException(GlobalErrorCode.ACCESS_DENIED);
+        }
         List<AdminReportedPostResponse> result = adminReportService.getReportedVideoBoard();
         return new BaseResponse<>(result);
     }
@@ -37,7 +43,10 @@ public class AdminReportController {
     @ApiResponses({
             @ApiResponse(responseCode = "200", description = "신고 내용 불러오기 성공")
     })
-    public BaseResponse<List<AdminReportDetailResponse>> getReportDetail(@RequestParam Long videoBoardId) {
+    public BaseResponse<List<AdminReportDetailResponse>> getReportDetail(@AuthenticationPrincipal AdminDetails adminDetails, @RequestParam Long videoBoardId) {
+        if (adminDetails == null) {
+            throw new GlobalBaseException(GlobalErrorCode.ACCESS_DENIED);
+        }
         List<AdminReportDetailResponse> reportDetail = adminReportService.getReportDetail(videoBoardId);
         return new BaseResponse<>(reportDetail);
     }
@@ -47,7 +56,10 @@ public class AdminReportController {
     @ApiResponses({
             @ApiResponse(responseCode = "200", description = "영상 숨김 해제 성공")
     })
-    public BaseResponse<?> cancelHiddenStatus(@RequestParam Long videoBoardId) {
+    public BaseResponse<?> cancelHiddenStatus(@AuthenticationPrincipal AdminDetails adminDetails, @RequestParam Long videoBoardId) {
+        if (adminDetails == null) {
+            throw new GlobalBaseException(GlobalErrorCode.ACCESS_DENIED);
+        }
         adminReportService.cancelHiddenStatus(videoBoardId);
         return new BaseResponse<>(GlobalErrorCode.SUCCESS);
     }
@@ -57,7 +69,10 @@ public class AdminReportController {
     @ApiResponses({
             @ApiResponse(responseCode = "200", description = "영상 삭제 성공")
     })
-    public BaseResponse<?> deleteVideo(@RequestParam Long videoBoardId) {
+    public BaseResponse<?> deleteVideo(@AuthenticationPrincipal AdminDetails adminDetails, @RequestParam Long videoBoardId) {
+        if (adminDetails == null) {
+            throw new GlobalBaseException(GlobalErrorCode.ACCESS_DENIED);
+        }
         adminReportService.deleteReportedVideo(videoBoardId);
         return new BaseResponse<>(GlobalErrorCode.SUCCESS);
     }
